remote_agents/sports_news_agent/pyproject.toml
[project]
name = "sports-news-agent"
version = "0.1.0"
description = "Multi-agent system with Azure AI Agents integration for routing and task delegation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "click>=8.2.1",
    "a2a-sdk==0.2.9",
    "azure-identity>=1.23.0",
    "azure-ai-agents>=1.2.0b3",
    "semantic-kernel>=1.36.1",
    "mcp>=1.9.4",
    "tiktoken>=0.5.0",
]